Load peak counts file into a SingleCellExperiment object

Load one 3-column counts.tab.gz file from polyApipe.py into a SingleCellExperiment for analysis.

load_peaks_counts_into_sce(counts_file, peak_info_file, batch = "",
  output, cell_prefix = batch, n_max = -1,
  min_reads_per_barcode = 100, missing_peaks_are_zero = TRUE, ...)

Arguments

counts_file
Tab-delimeted file of read counts per peak. The '<sample>.tab.gz' as output from polyApipe.py. Format: <peak> <cell> <count>

peak_info_file
GTF formatted peak file _specfically_ as output from polyApipe.py

batch
A name of the batch/sample. Important if multiple samples are going to be combined.

output
Where to save the sce object on disk using ssaveHDF5SummarizedExperiment. Should not exist. This directory will be created.

cell_prefix
A name to be prefixed to each cell's name. Important if multiple samples are going to be combined. (Default = batch)

n_max
Only read the first *n_max* lines of counts table. (Default = -1 (all))

min_reads_per_barcode
Only keep cells with at least this many reads. (Default = 100)

missing_peaks_are_zero
If there's nothing for a peak in the counts_file, assume there was zero reads. Needed to combine multiple samples. (Default = TRUE)

...
Extra options passed to saveHDF5SummarizedExperiment

See also

Examples

counts_file    <- system.file("extdata", "demo_dataset/demo_counts/SRR5259354_demo.tab.gz", package = "polyApiper")
peak_info_file <- system.file("extdata", "demo_dataset/demo_polyA_peaks.gff", package = "polyApiper")

sce <- load_peaks_counts_into_sce(counts_file    =counts_file1,
                                  peak_info_file = peak_info_file,
                                  output         = "demo",
                                  min_reads_per_barcode = 1)

#> Error in read.table(counts_file, sep = "\t", header = TRUE, as.is = TRUE,     colClasses = c("factor", "factor", "numeric"), nrows = n_max): object 'counts_file1' not found

# NB: min_reads_per_barcode=1 needed for test datatset due to its small size
#     use default (100) or something reasonable for real data!

if (FALSE) {

sce <- load_peaks_counts_into_sce(real_counts_file, "real_polyA_peaks.gff", output = "mysce")

sce <- load_peaks_counts_into_sce(counts_file    = "run1_knockout_counts.tab.gz",
                                  peak_info_file = "run1_knockout_polyA_peaks.gff",
                                  output         = "./knockout_rep1",
                                  batch          = "knockout_rep1")

# To reload from disk later 
library('HDF5Array')
library('SingleCellExperiment')
sce <- loadHDF5SummarizedExperiment("knockout_rep1/")

}